Implemented databases and save of every times in db. Screen's color changes when chronometer is running.

+++ app/src/main/java/com/example/cubelearner/chronometer/Counter.java
@@ -2,15 +2,15 @@ package com.example.cubelearner.chronometer;
 
 public class Counter {
 
-    private int value;
+    private long value;
 
     public Counter(){
         this.value = 0;
     }
-    public int getValue(){
+    public long getValue(){
         return value;
     }
-    public void setValue(int value) {
+    public void setValue(long value) {
         this.value = value;
     }
     public boolean increment(){
